+++ src/pynwb/device.py
@@ -1,6 +1,8 @@
 from hdmf.utils import docval, popargs
-from . import register_class, CORE_NAMESPACE
+
+from .globals import CORE_NAMESPACE
 from .core import NWBContainer
+from .utils import register_class
 
 
 @register_class('Device', CORE_NAMESPACE)
