+++ src/pynwb/ecephys.py
@@ -4,12 +4,13 @@ import warnings
 from hdmf.utils import docval, popargs, get_docval, popargs_to_dict
 from hdmf.data_utils import DataChunkIterator, assertEqualShape
 from hdmf.utils import get_data_shape
+from hdmf.common import DynamicTableRegion
 
-from . import register_class, CORE_NAMESPACE
+from .globals import CORE_NAMESPACE
 from .base import TimeSeries
 from .core import NWBContainer, NWBDataInterface, MultiContainerInterface
-from hdmf.common import DynamicTableRegion
 from .device import Device
+from .utils import register_class
 
 
 @register_class('ElectrodeGroup', CORE_NAMESPACE)
