+++ src/pynwb/misc.py
@@ -1,13 +1,14 @@
-import numpy as np
-from collections.abc import Iterable
 import warnings
+from collections.abc import Iterable
 from bisect import bisect_left, bisect_right
 
+import numpy as np
 from hdmf.utils import docval, getargs, popargs, popargs_to_dict, get_docval
-
-from . import register_class, CORE_NAMESPACE
-from .base import TimeSeries
 from hdmf.common import DynamicTable, DynamicTableRegion
+
+from .globals import CORE_NAMESPACE
+from .base import TimeSeries
+from .utils import register_class
 
 
 @register_class('AnnotationSeries', CORE_NAMESPACE)
